refactor and rename to predictWin
