+++ src/__tests__/predict-win.test.js
@@ -1,6 +1,6 @@
-import { rating, predict } from '..'
+import { rating, predictWin } from '..'
 
-describe('predict', () => {
+describe('predictWin', () => {
   const precision = 7
 
   const a1 = rating()
@@ -14,21 +14,21 @@ describe('predict', () => {
 
   it('predicts win outcome for two teams', () => {
     expect.assertions(2)
-    const [prob1, prob2] = predict([team1, team2])
+    const [prob1, prob2] = predictWin([team1, team2])
     expect(prob1).toBeCloseTo(0.34641823958165474, precision)
     expect(prob2).toBeCloseTo(0.6535817604183453, precision)
   })
 
   it('ignores rankings', () => {
     expect.assertions(1)
-    const p1 = predict([[a2], [b1], [b2]], { rank: [2, 1, 3] })
-    const p2 = predict([[a2], [b1], [b2]], { rank: [3, 2, 1] })
+    const p1 = predictWin([[a2], [b1], [b2]], { rank: [2, 1, 3] })
+    const p2 = predictWin([[a2], [b1], [b2]], { rank: [3, 2, 1] })
     expect(p1).toStrictEqual(p2)
   })
 
   it('predicts win outcome for multiple asymmetric teams', () => {
     expect.assertions(4)
-    const [prob1, prob2, prob3, prob4] = predict([team1, team2, [a2], [b2]])
+    const [prob1, prob2, prob3, prob4] = predictWin([team1, team2, [a2], [b2]])
     // should denom 6
     expect(prob1).toBeCloseTo(0.2613515941642222, precision)
     expect(prob2).toBeCloseTo(0.41117430943389155, precision)
@@ -38,7 +38,7 @@ describe('predict', () => {
 
   it('3 player newbie FFA', () => {
     expect.assertions(3)
-    const [prob1, prob2, prob3] = predict([[a1], [a1], [a1]])
+    const [prob1, prob2, prob3] = predictWin([[a1], [a1], [a1]])
     // denom 3
     expect(prob1).toBeCloseTo(0.333333333333, precision)
     expect(prob2).toBeCloseTo(0.333333333333, precision)
@@ -47,7 +47,7 @@ describe('predict', () => {
 
   it('4 player newbie FFA', () => {
     expect.assertions(4)
-    const [p1, p2, p3, p4] = predict([[a1], [a1], [a1], [a1]])
+    const [p1, p2, p3, p4] = predictWin([[a1], [a1], [a1], [a1]])
     // denom 6
     expect(p1).toBeCloseTo(0.25, precision)
     expect(p2).toBeCloseTo(0.25, precision)
@@ -61,7 +61,7 @@ describe('predict', () => {
     const r2 = rating({ mu: 2, sigma: 0.1 })
     const r3 = rating({ mu: 3, sigma: 0.1 })
     const r4 = rating({ mu: 4, sigma: 0.1 })
-    const [p1, p2, p3, p4] = predict([[r1], [r2], [r3], [r4]])
+    const [p1, p2, p3, p4] = predictWin([[r1], [r2], [r3], [r4]])
     expect(p1).toBeCloseTo(0.2028051110543726, precision)
     expect(p2).toBeCloseTo(0.23419421333676907, precision)
     expect(p3).toBeCloseTo(0.2658057866632309, precision)
@@ -71,7 +71,7 @@ describe('predict', () => {
   it('5 player newbie FFA', () => {
     expect.assertions(5)
     // denom 10
-    const [p1, p2, p3, p4, p5] = predict([[a1], [a1], [a1], [a1], [a1]])
+    const [p1, p2, p3, p4, p5] = predictWin([[a1], [a1], [a1], [a1], [a1]])
     expect(p1).toBeCloseTo(0.2, precision)
     expect(p2).toBeCloseTo(0.2, precision)
     expect(p3).toBeCloseTo(0.2, precision)
@@ -81,7 +81,7 @@ describe('predict', () => {
 
   it('5 player FFA with an impostor', () => {
     expect.assertions(5)
-    const [p1, p2, p3, p4, p5] = predict([[a1], [a1], [a1], [a2], [a1]])
+    const [p1, p2, p3, p4, p5] = predictWin([[a1], [a1], [a1], [a2], [a1]])
     expect(p1).toBeCloseTo(0.196037416522638, precision)
     expect(p2).toBeCloseTo(0.196037416522638, precision)
     expect(p3).toBeCloseTo(0.196037416522638, precision)
